requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
boto3==1.36.21
    # via -r requirements.in
botocore==1.36.21
    # via
    #   boto3
    #   s3transfer
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
django==5.1.1
    # via
    #   -r requirements.in
    #   django-storages
    #   django-taggit
    #   django-use-email-as-username
    #   djangorestframework
django-storages==1.14.5
    # via -r requirements.in
django-taggit==6.1.0
    # via -r requirements.in
django-use-email-as-username==1.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
faker==28.0.0
    # via -r requirements.in
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
pillow==10.4.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   faker
python-decouple==3.8
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
s3transfer==0.11.2
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
tzdata==2024.2
    # via django
urllib3==2.3.0
    # via
    #   botocore
    #   requests
whitenoise==6.7.0
    # via -r requirements.in
